Rename ASCII board module, add board regression
