+++ src/dh_workspace/__init__.py
@@ -16,7 +16,7 @@ from .core.backend.pieces import (
 from .core.backend.match import Match
 from .utils.config import CONFIG, Config, configure
 from .utils.logger import logger
-from .frontend.ascii_board import draw_board, draw_empty_board, save_board
+from .frontend.unicode_board import draw_board, draw_empty_board, save_board
 
 __all__ = [
     "Chessboard",
