+++ tests/resources/empty_board.txt
@@ -1,9 +1,17 @@
-_ _ _ _ _ _ _ _
-|_|_|_|_|_|_|_|_|
-|_|_|_|_|_|_|_|_|
-|_|_|_|_|_|_|_|_|
-|_|_|_|_|_|_|_|_|
-|_|_|_|_|_|_|_|_|
-|_|_|_|_|_|_|_|_|
-|_|_|_|_|_|_|_|_|
-|_|_|_|_|_|_|_|_|
+┏━━━┳━━━┳━━━┳━━━┳━━━┳━━━┳━━━┳━━━┓
+┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
+┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
+┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
+┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
+┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
+┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
+┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
+┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
+┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
+┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
+┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
+┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
+┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
+┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
+┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
+┗━━━┻━━━┻━━━┻━━━┻━━━┻━━━┻━━━┻━━━┛
