+++ tests/resources/pawn_knight_moves.txt
@@ -1,17 +1,17 @@
 ┏━━━┳━━━┳━━━┳━━━┳━━━┳━━━┳━━━┳━━━┓
-┃ ♜ ┃ ♞ ┃ ♝ ┃ ♛ ┃ ♚ ┃ ♝ ┃ ♞ ┃ ♜ ┃
+┃ ♜ ┃   ┃ ♝ ┃ ♛ ┃ ♚ ┃ ♝ ┃ ♞ ┃ ♜ ┃
 ┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
-┃ ♟ ┃ ♟ ┃ ♟ ┃ ♟ ┃ ♟ ┃ ♟ ┃ ♟ ┃ ♟ ┃
+┃   ┃ ♟ ┃ ♟ ┃ ♟ ┃ ♟ ┃ ♟ ┃ ♟ ┃ ♟ ┃
 ┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
-┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
-┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
-┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
+┃ ♟ ┃   ┃ ♞ ┃   ┃   ┃   ┃   ┃   ┃
 ┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
 ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
 ┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
 ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃   ┃
 ┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
-┃ ♙ ┃ ♙ ┃ ♙ ┃ ♙ ┃ ♙ ┃ ♙ ┃ ♙ ┃ ♙ ┃
+┃ ♙ ┃   ┃ ♘ ┃   ┃   ┃   ┃   ┃   ┃
 ┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
-┃ ♖ ┃ ♘ ┃ ♗ ┃ ♕ ┃ ♔ ┃ ♗ ┃ ♘ ┃ ♖ ┃
+┃   ┃ ♙ ┃ ♙ ┃ ♙ ┃ ♙ ┃ ♙ ┃ ♙ ┃ ♙ ┃
+┣━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━╋━━━┫
+┃ ♖ ┃   ┃ ♗ ┃ ♕ ┃ ♔ ┃ ♗ ┃ ♘ ┃ ♖ ┃
 ┗━━━┻━━━┻━━━┻━━━┻━━━┻━━━┻━━━┻━━━┛
